Multiplayer

+++ SSPSolution/SSPSolution/SoundHandler.cpp
@@ -431,10 +431,11 @@ void SoundHandler::DropSounds()
 	}
 }
 
+
 irrklang::ISound* SoundHandler::PlaySound2D(Sounds2D soundEnum, bool loop, bool track)
 {
 	//Check if the enum will fit as an index
-	if (soundEnum < this->m_sounds2D.size() && soundEnum != Sounds2D::NO_SOUND2D)
+	if (soundEnum < (int)this->m_sounds2D.size() && soundEnum != Sounds2D::NO_SOUND2D)
 	{
 		irrklang::ISoundSource* sp = this->m_sounds2D.at(soundEnum);
 		irrklang::ISound* newActiveSound = this->m_soundEngine->play2D(sp, loop, false, true);
@@ -460,7 +461,7 @@ irrklang::ISound* SoundHandler::PlaySound2D(Sounds2D soundEnum, bool loop, bool 
 irrklang::ISound* SoundHandler::PlaySound3D(Sounds3D soundEnum, DirectX::XMFLOAT3 pos, bool loop, bool track)
 {
 	//Check if the enum will fit as an index
-	if (soundEnum < this->m_sounds3D.size() && soundEnum != Sounds3D::NO_SOUND3D)
+	if (soundEnum < (int)this->m_sounds3D.size() && soundEnum != Sounds3D::NO_SOUND3D)
 	{
 		irrklang::ISoundSource* sp = this->m_sounds3D.at(soundEnum);
 		irrklang::vec3d<float> pos(pos.x, pos.y, pos.z);
@@ -487,8 +488,8 @@ irrklang::ISound* SoundHandler::PlaySound3D(Sounds3D soundEnum, DirectX::XMFLOAT
 irrklang::ISound * SoundHandler::PlayRandomSound2D(Sounds2D start_soundEnum, Sounds2D end_soundEnum, bool loop, bool track)
 {
 	//Check if the enum will fit as an index
-	if (start_soundEnum < this->m_sounds2D.size() && start_soundEnum != Sounds2D::NO_SOUND2D &&
-		end_soundEnum < this->m_sounds2D.size() && end_soundEnum != Sounds2D::NO_SOUND2D)
+	if (start_soundEnum < (int)this->m_sounds2D.size() && start_soundEnum != Sounds2D::NO_SOUND2D &&
+		end_soundEnum < (int)this->m_sounds2D.size() && end_soundEnum != Sounds2D::NO_SOUND2D)
 	{
 
 		int randomSoundID = (rand() % (end_soundEnum - start_soundEnum)) + start_soundEnum;	//Get a random sound between the two defined sounds
@@ -517,8 +518,8 @@ irrklang::ISound * SoundHandler::PlayRandomSound2D(Sounds2D start_soundEnum, Sou
 irrklang::ISound * SoundHandler::PlayRandomSound3D(Sounds3D start_soundEnum, Sounds3D end_soundEnum, DirectX::XMFLOAT3 pos, bool loop, bool track)
 {
 	//Check if the enum will fit as an index
-	if (start_soundEnum < this->m_sounds3D.size() && start_soundEnum != Sounds3D::NO_SOUND3D &&
-		end_soundEnum < this->m_sounds3D.size() && end_soundEnum != Sounds3D::NO_SOUND3D)
+	if (start_soundEnum < (int)this->m_sounds3D.size() && start_soundEnum != Sounds3D::NO_SOUND3D &&
+		end_soundEnum < (int)this->m_sounds3D.size() && end_soundEnum != Sounds3D::NO_SOUND3D)
 	{
 		int randomSoundID = (rand() % (end_soundEnum - start_soundEnum)) + start_soundEnum;	//Get a random sound between the two defined sounds
